requirements.txt
requests
beautifulsoup4
google-cloud-speech
youtube-dl
pydub